// moonbowcl/src/app/model/dto/attr/AttrCrmConfR2U.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class AttrCrmConfR2U {
    riepilogo: string;
    dettaglio: string;
    tipologia: string;
    urgenza: string;
    impatto: string;
    include_allegati: boolean;
    assegnatario: string;
    emailForFailure: string;
    oggettoEmail: string;
    bodyEmail: string;
    //codiceFiscale: string;
    nome: string;
    cognome: string;
    email: string;
    phone: string;
    mobile: string;
    companyName: string;
    //companyId: string;
    cfi: string;
    mexNewUser: string;
    livello1: string;
    livello2: string;
    livello3: string;
}